refactor(transform): isolate OXC scope traversal
